day 09

+++ src/main/java/uk/co/swilson/advent/Solver.java
@@ -1,6 +1,6 @@
 package uk.co.swilson.advent;
 
 public interface Solver {
-    int solvePartOne(String input);
-    int solvePartTwo(String input);
+    long solvePartOne(String input);
+    long solvePartTwo(String input);
 }
